+++ src/mos6502/mos6502_sty.c
@@ -16,18 +16,18 @@
 size_t
 mos6502_sty_abs8(int8_t value)
 {
-	return mos6502_write(0x84) + mos6502_write(value);
+	return mos6502_write((int8_t) 0x84) + mos6502_write(value);
 }
 
 size_t
 mos6502_sty_abs8_rx(int8_t value)
 {
-	return mos6502_write(0x94) + mos6502_write(value);
+	return mos6502_write((int8_t) 0x94) + mos6502_write(value);
 }
 
 size_t
 mos6502_sty_abs16(int16_t value)
 {
-	return mos6502_write(0x8c) + mos6502_imm16(value);
+	return mos6502_write((int8_t) 0x8c) + mos6502_imm16(value);
 }
