+++ src/z80/z80_xor.c
@@ -17,27 +17,27 @@ size_t
 z80_xor_r8(Z80_REG reg)
 {
 	if (z80_idx(reg))
-		return z80_i(reg) + z80_b(0xac, reg);
+		return z80_i(reg) + z80_b((int8_t) 0xac, reg);
 	else
-		return z80_rrr(0xa8, 0, reg);
+		return z80_rrr((int8_t) 0xa8, 0, reg);
 }
 
 size_t
 z80_xor_imm8(int8_t value)
 {
-	return z80_write(0xee) + z80_write(value);
+	return z80_write((int8_t) 0xee) + z80_write(value);
 }
 
 size_t
 z80_xor_rhl()
 {
-	return z80_write(0xae);
+	return z80_write((int8_t) 0xae);
 }
 
 size_t
 z80_xor_ir16_off8(Z80_REG reg,
                   int8_t off)
 {
-	return z80_i(reg) + z80_write(0xae) + z80_write(off);
+	return z80_i(reg) + z80_write((int8_t) 0xae) + z80_write(off);
 }
